Relocate pubsub
