+++ scio-core/src/main/scala/com/spotify/scio/ScioContext.scala
@@ -754,123 +754,6 @@ class ScioContext private[scio] (
   def datastore(projectId: String, query: Query, namespace: String = null): SCollection[Entity] =
     this.read(DatastoreIO(projectId))(DatastoreIO.ReadParam(query, namespace))
 
-  // This method needs to be removed once pubsubSubscription and pubsubTopic are deleted
-  private def pubsubIn[T: ClassTag: Coder](
-    isSubscription: Boolean,
-    name: String,
-    idAttribute: String,
-    timestampAttribute: String
-  ): SCollection[T] = {
-    val io = PubsubIO[T](name, idAttribute, timestampAttribute)
-    this.read(io)(PubsubIO.ReadParam(isSubscription))
-  }
-
-  /**
-   * Get an SCollection for a Pub/Sub subscription.
-   * @group input
-   */
-  @deprecated(
-    """
-    |  This method has been deprecated. Use one of the following IOs instead:
-    |    - PubsubIO.string
-    |    - PubsubIO.avro
-    |    - PubsubIO.proto
-    |    - PubsubIO.pubsub
-    |    - PubsubIO.coder
-    |
-    |  For example:
-    |     sc.read(PubsubIO.string(sub, idAttribute, timestampAttribute))(
-    |       PubsubIO.ReadParam(PubsubIO.Subscription)
-    |     )
-    """.stripMargin,
-    since = "0.10.0"
-  )
-  def pubsubSubscription[T: ClassTag: Coder](
-    sub: String,
-    idAttribute: String = null,
-    timestampAttribute: String = null
-  ): SCollection[T] =
-    pubsubIn(isSubscription = true, sub, idAttribute, timestampAttribute)
-
-  /**
-   * Get an SCollection for a Pub/Sub topic.
-   * @group input
-   */
-  @deprecated(
-    """
-    |  This method has been deprecated. Use one of the following IOs instead:
-    |    - PubsubIO.string
-    |    - PubsubIO.avro
-    |    - PubsubIO.proto
-    |    - PubsubIO.pubsub
-    |    - PubsubIO.coder
-    |
-    |  For example:
-    |     sc.read(PubsubIO.string(sub, idAttribute, timestampAttribute))(
-    |       PubsubIO.ReadParam(PubsubIO.Topic)
-    |     )
-    """.stripMargin,
-    since = "0.10.0"
-  )
-  def pubsubTopic[T: ClassTag: Coder](
-    topic: String,
-    idAttribute: String = null,
-    timestampAttribute: String = null
-  ): SCollection[T] =
-    pubsubIn(isSubscription = false, topic, idAttribute, timestampAttribute)
-
-  private def pubsubInWithAttributes[T: ClassTag: Coder](
-    isSubscription: Boolean,
-    name: String,
-    idAttribute: String,
-    timestampAttribute: String
-  ): SCollection[(T, Map[String, String])] = {
-    val io = PubsubIO.withAttributes[T](name, idAttribute, timestampAttribute)
-    this.read(io)(PubsubIO.ReadParam(isSubscription))
-  }
-
-  /**
-   * Get an SCollection for a Pub/Sub subscription that includes message attributes.
-   * @group input
-   */
-  @deprecated(
-    """
-    |  This method has been deprecated. Use PubsubIO.withAttributes instead.
-    |  For example:
-    |     sc.read(PubsubIO.withAttributes(sub, idAttribute, timestampAttribute))(
-    |       PubsubIO.ReadParam(PubsubIO.Subscription)
-    |     )
-    """.stripMargin,
-    since = "0.10.0"
-  )
-  def pubsubSubscriptionWithAttributes[T: ClassTag: Coder](
-    sub: String,
-    idAttribute: String = null,
-    timestampAttribute: String = null
-  ): SCollection[(T, Map[String, String])] =
-    pubsubInWithAttributes[T](isSubscription = true, sub, idAttribute, timestampAttribute)
-
-  /**
-   * Get an SCollection for a Pub/Sub topic that includes message attributes.
-   * @group input
-   */
-  @deprecated(
-    """
-    |  This method has been deprecated. Use PubsubIO.withAttributes instead.
-    |  For example:
-    |     sc.read(PubsubIO.withAttributes(sub, idAttribute, timestampAttribute))(
-    |       PubsubIO.ReadParam(PubsubIO.Topic)
-    |     )
-    """.stripMargin,
-    since = "0.10.0"
-  )
-  def pubsubTopicWithAttributes[T: ClassTag: Coder](
-    topic: String,
-    idAttribute: String = null,
-    timestampAttribute: String = null
-  ): SCollection[(T, Map[String, String])] =
-    pubsubInWithAttributes[T](isSubscription = false, topic, idAttribute, timestampAttribute)
-
   /**
    * Get an SCollection for a text file.
    * @group input
