+++ scio-google-cloud-platform/src/main/scala/com/spotify/scio/pubsub/PubsubCoders.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2019 Spotify AB.
+ * Copyright 2020 Spotify AB.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,16 @@
  * specific language governing permissions and limitations
  * under the License.
  */
+package com.spotify.scio.pubsub
 
-package com.spotify.scio.io
+import com.spotify.scio.coders.Coder
+import org.apache.beam.sdk.io.gcp.pubsub.{PubsubMessage, PubsubMessageWithAttributesCoder}
 
-import com.spotify.scio.testing.PipelineSpec
-import com.google.protobuf.Message
-import org.apache.beam.sdk.io.gcp.{pubsub => beam}
+trait PubsubCoders {
 
-class PubsubIOTest extends PipelineSpec {
-  "PubsubIO" should "#2582: not throw an ClassCastException when created" in {
-    PubsubIO[String]("String IO")
-    PubsubIO[com.spotify.scio.avro.Account]("SpecificRecordBase IO")
-    PubsubIO[Message]("Message IO")
-    PubsubIO[beam.PubsubMessage]("Message IO")
-  }
+  implicit def messageCoder: Coder[PubsubMessage] =
+    Coder.beam(PubsubMessageWithAttributesCoder.of())
+
 }
+
+object coders extends PubsubCoders
