+++ scio-test/src/test/scala/com/spotify/scio/testing/JobTestTest.scala
@@ -31,7 +31,6 @@ import org.apache.avro.generic.GenericRecord
 import org.apache.beam.sdk.Pipeline.PipelineExecutionException
 import org.apache.beam.sdk.metrics.DistributionResult
 import org.apache.beam.sdk.{io => beam}
-import org.joda.time.Instant
 import org.scalatest.exceptions.TestFailedException
 
 import scala.io.Source
@@ -92,30 +91,6 @@ object DatastoreJob {
     val (sc, args) = ContextAndArgs(cmdlineArgs)
     sc.datastore(args("input"), null, null)
       .saveAsDatastore(args("output"))
-    sc.run()
-    ()
-  }
-}
-
-object PubsubJob {
-  def main(cmdlineArgs: Array[String]): Unit = {
-    val (sc, args) = ContextAndArgs(cmdlineArgs)
-    sc.pubsubTopic[String](args("input"))
-      .map(_ + "X")
-      .saveAsPubsub(args("output"))
-    sc.run()
-    ()
-  }
-}
-
-object PubsubWithAttributesJob {
-  val timestampAttribute = "tsAttribute"
-
-  def main(cmdlineArgs: Array[String]): Unit = {
-    val (sc, args) = ContextAndArgs(cmdlineArgs)
-    sc.pubsubTopicWithAttributes[String](args("input"), timestampAttribute = timestampAttribute)
-      .map(kv => (kv._1 + "X", kv._2))
-      .saveAsPubsubWithAttributes[String](args("output"), timestampAttribute = timestampAttribute)
     sc.run()
     ()
   }
@@ -361,101 +336,6 @@ class JobTestTest extends PipelineSpec {
     }
     an[AssertionError] should be thrownBy {
       testDatastore((1 to 4).map(newEntity))
-    }
-  }
-
-  def testPubsubJob(xs: String*): Unit =
-    JobTest[PubsubJob.type]
-      .args("--input=in", "--output=out")
-      .input(PubsubIO[String]("in"), Seq("a", "b", "c"))
-      .output(PubsubIO[String]("out"))(coll => coll should containInAnyOrder(xs))
-      .run()
-
-  it should "pass correct PubsubIO" in {
-    testPubsubJob("aX", "bX", "cX")
-  }
-
-  it should "fail incorrect PubsubIO" in {
-    an[AssertionError] should be thrownBy { testPubsubJob("aX", "bX") }
-    an[AssertionError] should be thrownBy {
-      testPubsubJob("aX", "bX", "cX", "dX")
-    }
-  }
-
-  def testPubsubJobWithTestStreamInput(xs: String*): Unit =
-    JobTest[PubsubJob.type]
-      .args("--input=in", "--output=out")
-      .inputStream(
-        PubsubIO[String]("in"),
-        testStreamOf[String].addElements("a", "b", "c").advanceWatermarkToInfinity()
-      )
-      .output(PubsubIO[String]("out"))(coll => coll should containInAnyOrder(xs))
-      .run()
-
-  it should "pass correct PubsubIO with TestStream input" in {
-    testPubsubJobWithTestStreamInput("aX", "bX", "cX")
-  }
-
-  it should "fail incorrect PubsubIO with TestStream input" in {
-    an[AssertionError] should be thrownBy testPubsubJobWithTestStreamInput("aX", "bX")
-    an[AssertionError] should be thrownBy testPubsubJobWithTestStreamInput("aX", "bX", "cX", "dX")
-  }
-
-  def testPubsubWithAttributesJob(timestampAttribute: Map[String, String], xs: String*): Unit = {
-    type M = Map[String, String]
-    val m = Map("a" -> "1", "b" -> "2", "c" -> "3") ++ timestampAttribute
-    JobTest[PubsubWithAttributesJob.type]
-      .args("--input=in", "--output=out")
-      .input(
-        PubsubIO[(String, M)]("in", null, PubsubWithAttributesJob.timestampAttribute),
-        Seq("a", "b", "c").map((_, m))
-      )
-      .output(
-        PubsubIO[(String, M)]("out", null, PubsubWithAttributesJob.timestampAttribute)
-      )(coll => coll should containInAnyOrder(xs.map((_, m))))
-      .run()
-  }
-
-  it should "pass correct PubsubIO with attributes" in {
-    testPubsubWithAttributesJob(
-      Map(PubsubWithAttributesJob.timestampAttribute -> new Instant().toString),
-      "aX",
-      "bX",
-      "cX"
-    )
-  }
-
-  it should "fail incorrect PubsubIO with attributes" in {
-    an[AssertionError] should be thrownBy {
-      testPubsubWithAttributesJob(
-        Map(PubsubWithAttributesJob.timestampAttribute -> new Instant().toString),
-        "aX",
-        "bX"
-      )
-    }
-    an[AssertionError] should be thrownBy {
-      testPubsubWithAttributesJob(
-        Map(PubsubWithAttributesJob.timestampAttribute -> new Instant().toString),
-        "aX",
-        "bX",
-        "cX",
-        "dX"
-      )
-    }
-  }
-
-  it should "fail PubsubIO with invalid or missing timestamp attribute" in {
-    an[PipelineExecutionException] should be thrownBy {
-      testPubsubWithAttributesJob(
-        Map(PubsubWithAttributesJob.timestampAttribute -> "invalidTimestamp"),
-        "aX",
-        "bX",
-        "cX"
-      )
-    }
-
-    an[PipelineExecutionException] should be thrownBy {
-      testPubsubWithAttributesJob(timestampAttribute = Map(), xs = "aX", "bX", "cX")
     }
   }
